work in progress
Former-commit-id: 08090ec0251bb3bfd1cb2fae2f66c805303b89ed
Former-commit-id: cade495fafd6ea3c4b6bdf862a3cd934f3c5c19f

+++ src/tree_test.c
@@ -50,13 +50,13 @@ typedef struct tree_debug_node_ {
 
 static tree_debug_node tree_debug_nodes[TREE_DEBUG_NUM_NODES];
 
-static int32_t tree_debug_compare_func(const tree_node *, const tree_node *);
+static int8_t tree_debug_compare_func(const tree_node *, const tree_node *);
 
 tree_root tree_debug_root = {
     tree_debug_compare_func
 };
 
-static int32_t tree_debug_compare_func (const tree_node *a, const tree_node *b)
+static int8_t tree_debug_compare_func (const tree_node *a, const tree_node *b)
 {
     tree_debug_node *A = (typeof(A))a;
     tree_debug_node *B = (typeof(B))b;
